requirements.txt
# Data Processing
pandas>=2.2.0
numpy>=1.26.0,<2.2  # Constraint for imbalanced-learn compatibility

# Machine Learning
scikit-learn>=1.4.0
xgboost>=2.1.0
imbalanced-learn>=0.12.0

# Data Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Dashboard
streamlit>=1.29.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
faker>=20.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Model Interpretability
shap>=0.44.0

# Data Generation
openpyxl>=3.1.0
